requirements.txt
pyzmq==16.0.2
ipython==6.2.1
pytest==3.2.3
